Refactored to make main.py cleaner
