+++ model/framework/code/image_dataloader.py
@@ -1,3 +1,8 @@
+import os
+import glob
+
+import torch
+import torchvision
 from PIL import Image
 from rdkit import Chem
 from rdkit.Chem import Draw
@@ -9,9 +14,17 @@ from rdkit.Chem import Descriptors
 import numpy as np
 import torchvision.transforms as transforms
 
+root = os.path.dirname(os.path.abspath(__file__))
+checkpoints_dir = os.path.abspath(os.path.join(root, "..", "..", "checkpoints"))
+ckpts = glob.glob(f"{checkpoints_dir}/*.pth")
+gpcr_assays = [os.path.split(pt)[-1].split(".")[0] for pt in ckpts]
+
+
 IMG_SIZE = 224
 ORGANIC_ATOM_SET = set([5, 6, 7, 8, 9, 15, 16, 17, 35, 53])
 REMOVER = SaltRemover()
+
+DEVICE = 'cpu'
 
 
 class ImageData:
@@ -48,3 +61,39 @@ def smiles_to_image(smi, size=IMG_SIZE, savePath=None):
     except Exception as err:
         print(err)
         return None
+
+
+def load_models():
+    models = {}
+
+    for idx, assay in enumerate(gpcr_assays):
+        mdl = torchvision.models.resnet18(weights=None)
+        mdl.fc = torch.nn.Linear(mdl.fc.in_features, 1)
+        checkpoint = torch.load(ckpts[idx], map_location=torch.device(DEVICE))
+        mdl.load_state_dict(checkpoint)
+        models.update({assay: mdl})
+
+    return models
+
+
+# Get predictions
+def get_predictions(smiles):
+    models = load_models()
+
+    outputs = []
+
+    img_processor = ImageData()
+    for idx, smi in enumerate(smiles):
+        per_row_preds = []
+        path = f"{os.getcwd()}/{idx}.png"
+        smiles_to_image(smi, savePath=path)
+        img_tensor = img_processor.get_image(path).to(DEVICE)
+        for assay in gpcr_assays:
+            assay_mdl = models[assay]
+            assay_mdl.eval()
+            with torch.no_grad():
+                pred = assay_mdl(img_tensor)
+            per_row_preds.append(pred.item())
+        outputs.append(per_row_preds)
+        os.remove(path)
+    return gpcr_assays,outputs
